requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
cffi==1.16.0
    # via cryptography
crispy-bootstrap5==2023.10
    # via -r requirements.in
cryptography==41.0.7
    # via -r requirements.in
django==4.2.7
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-ckeditor
    #   django-crispy-forms
    #   django-extensions
    #   django-js-asset
django-ckeditor==6.7.0
    # via -r requirements.in
django-cleanup==8.0.0
    # via -r requirements.in
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-extensions==3.2.3
    # via -r requirements.in
django-js-asset==2.1.0
    # via django-ckeditor
graphviz==0.20.1
    # via -r requirements.in
markupsafe==2.1.3
    # via werkzeug
pillow==10.1.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.4.4
    # via django
werkzeug==3.0.1
    # via -r requirements.in
